requirements.test.txt
# Minimal requirements untuk testing API refactoring
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Redis untuk caching (optional)
redis==5.0.1

# Untuk testing tanpa full dependencies
aiosqlite==0.19.0
